Improve SubPatch stability
Still destroy its wires if inlet/outlet count changes
